Migrate token type from uint16 to int32 for Qwen3 vocab (151K), fix 4B config
- Change all token types from uint16 to int32 across 15 files
- Qwen3 vocab_size=151936 overflows uint16 (max 65535)
- loadTokens reads uint16 from disk, returns []int32
- EvalLogits, EvalNextToken, Generate, etc. all use int32 tokens
- Fix Qwen3_4BConfig preset: hidden=9728 (was incorrect 6912)
- Update cmd/generate with HF tokenizer integration and proper EOS (151645)
- FP16 uint16 usage in IOSurface/MIL paths correctly unchanged

Tested: Qwen3-0.6B runs at 46 tok/s with proper tokenized output.

+++ ane/stories/cpu.go
@@ -38,7 +38,7 @@ func parallelFor(n int, fn func(start, end int)) {
 	wg.Wait()
 }
 
-func EmbedLookup(out, embed []float32, tokens []uint16, dim, seq int) {
+func EmbedLookup(out, embed []float32, tokens []int32, dim, seq int) {
 	if dim <= 0 || seq <= 0 {
 		return
 	}
@@ -71,7 +71,7 @@ func EmbedLookup(out, embed []float32, tokens []uint16, dim, seq int) {
 	})
 }
 
-func EmbedBackward(dEmbed, dx []float32, tokens []uint16, dim, seq int) {
+func EmbedBackward(dEmbed, dx []float32, tokens []int32, dim, seq int) {
 	if dim <= 0 || seq <= 0 {
 		return
 	}
@@ -170,7 +170,7 @@ func rmsNormBackwardRange(dx, dw, dy, x, w []float32, d, s, start, end int) {
 	}
 }
 
-func CrossEntropyLoss(dLogits, logits []float32, targets []uint16, v, s int) float32 {
+func CrossEntropyLoss(dLogits, logits []float32, targets []int32, v, s int) float32 {
 	if v <= 0 || s <= 0 {
 		return 0
 	}
